requirements.txt
# pip3 install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu128
tqdm
pyyaml
PyAudioWPatch
openai-whisper
pydub